[BUGFIX] Make some DB fields visually smaller

+++ ext_tables.sql
@@ -3,11 +3,11 @@
 #
 CREATE TABLE fe_users (
 	full_salutation varchar(255) DEFAULT '' NOT NULL,
-	gender int(11) unsigned DEFAULT '0' NOT NULL,
+	gender tinyint(2) unsigned DEFAULT '0' NOT NULL,
 	date_of_birth int(11) DEFAULT '0' NOT NULL,
 	zone varchar(45) DEFAULT '' NOT NULL,
-	privacy tinyint(4) unsigned DEFAULT '0' NOT NULL,
+	privacy tinyint(1) unsigned DEFAULT '0' NOT NULL,
 	terms_acknowledged tinyint(1) unsigned DEFAULT '0' NOT NULL,
-	status int(11) unsigned DEFAULT '0' NOT NULL,
+	status tinyint(1) unsigned DEFAULT '0' NOT NULL,
 	comments text
 );
